DBZ-252 refactor parser listener
